now casting deltaX/Y to double when taking floor in movePlayer()"

+++ A2/player.c
@@ -12,15 +12,14 @@
 //   A   M
 //  <x> <X>
 //   V   V
-char* PLAYER_BODY[PLAYER_BODY_ANIM_TILES][PLAYER_HEIGHT] = 
-{
-  {" A ",
-   "<x>",
-   " V "},
-  {" M ",
-   "<X>",
-   " V "}
-};
+char *PLAYER_BODY[PLAYER_BODY_ANIM_TILES][PLAYER_HEIGHT] =
+    {
+        {" A ",
+         "<x>",
+         " V "},
+        {" M ",
+         "<X>",
+         " V "}};
 
 int initPlayer()
 {
@@ -38,13 +37,13 @@ int movePlayer(int deltaX, int deltaY)
         return 0;
     }
 
-    int newPosX = PLAYER_POS_X + floor(deltaX);
+    int newPosX = PLAYER_POS_X + (int) floor((double)deltaX);
     if (newPosX <= GAME_COLS && newPosX >= 0)
     {
         PLAYER_POS_X = newPosX;
     }
 
-    int newPosY = PLAYER_POS_Y + floor(deltaY);
+    int newPosY = PLAYER_POS_Y + (int )floor((double)deltaY);
     if (newPosY <= GAME_ROWS && newPosY > 16)
     {
         PLAYER_POS_Y = newPosY;
